refactor(store): move helper to store
since helper do exports only store-related functions, they should be under store dir
