+++ src/interface/statusioStoreInterface.js
@@ -2,7 +2,7 @@ import PlatformStatus from './storeInterface';
 import {
   createRawStatusStore,
   createSummarizedStatusStore,
-} from '../helper/statusioHelper';
+} from '../store/statusioStore';
 
 class StatusioStore extends PlatformStatus {
   constructor(platform, statusId) {
